Refactor CLI into dedicated modules
