+++ projects/04-llm-adapter-shadow/src/llm_adapter/cli/__init__.py
@@ -1,0 +1,13 @@
+from __future__ import annotations
+
+from .args import parse_args
+from .config import build_runner_config
+from .io import prepare_execution
+from .runner import _format_output, main
+
+__all__ = [
+    "parse_args",
+    "build_runner_config",
+    "prepare_execution",
+    "main",
+]
